Remove diff generation during import
