+++ server/dbutil/dbutil_test.go
@@ -1,7 +1,6 @@
 package dbutil
 
 import (
-	"io/ioutil"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -21,41 +20,6 @@ func (suite *DBUtilSuite) TestMD5() {
 	for _, hash := range hashes {
 		assert.Equal(suite.T(), hash[1], md5hash(hash[0]))
 	}
-}
-
-func readFile(filename string) (string, error) {
-	data, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return "", err
-	}
-
-	return string(data), nil
-}
-
-func (suite *DBUtilSuite) TestDiff() {
-	assert := suite.Assert()
-
-	a, err := readFile("./testdata/a.txt")
-	assert.NoError(err)
-	b, err := readFile("./testdata/b.txt")
-	assert.NoError(err)
-	c, err := readFile("./testdata/c.txt")
-	assert.NoError(err)
-
-	ab, err := readFile("./testdata/ab.diff")
-	assert.NoError(err)
-	ac, err := readFile("./testdata/ac.diff")
-	assert.NoError(err)
-
-	abDiff, err := diff("a.txt", "b.txt", a, b)
-
-	assert.NoError(err)
-	assert.Equal(ab, abDiff)
-
-	acDiff, err := diff("a.txt", "c.txt", a, c)
-
-	assert.NoError(err)
-	assert.Equal(ac, acDiff)
 }
 
 func TestDBUtil(t *testing.T) {
