+++ server/service/testdata/ab.diff
@@ -1,7 +1,9 @@
 --- a.txt
 +++ b.txt
-@@ -1,3 +1,4 @@
+@@ -1,5 +1,6 @@
 +zero
  one
  two
  three
+ four
+ five
